Examination Types  - Allow for multiple final/midterm examination types in one academic session.   - Examination types are now editable in the application.   - Only examination types that have examination periods defined are visible within an academic session.

+++ JavaSource/org/unitime/timetable/action/ajax/DistributionPrefsAjax.java
@@ -96,7 +96,7 @@ public class DistributionPrefsAjax extends Action {
         } else if ("distType".equals(request.getParameter("type"))) {
             computePreferenceLevels(request.getParameter("id"),out);
         } else if ("exam".equals(request.getParameter("type"))) {
-            coumputeExams(request.getParameter("id"), Integer.valueOf(request.getParameter("examType")),out);
+            coumputeExams(request.getParameter("id"), Long.valueOf(request.getParameter("examType")),out);
         }
     }
     
@@ -184,7 +184,7 @@ public class DistributionPrefsAjax extends Action {
         }
     }
    
-    protected void coumputeExams(String courseOfferingId, Integer examType, PrintWriter out) throws Exception {
+    protected void coumputeExams(String courseOfferingId, Long examType, PrintWriter out) throws Exception {
         if (courseOfferingId==null || courseOfferingId.length()==0 || courseOfferingId.equals(Preference.BLANK_PREF_VALUE)) return;
         TreeSet exams = new TreeSet(Exam.findExamsOfCourseOffering(Long.valueOf(courseOfferingId),examType));
         if (exams.size()>1 || exams.isEmpty())
